+++ src/components/Story/StorySlider/StorySliderView.tsx
@@ -1,12 +1,45 @@
 import React from "react";
+import Link from "next/link";
 import { usePlayer } from "~/components/Player";
-import { Story, useNowPlayingQuery, useTrackQuery } from "~/graphql/gql.gen";
 import PlayerView from "~/components/Player/PlayerView";
 import StoryNav from "~/components/Story/StoryNav";
+import {
+  Story,
+  useNowPlayingQuery,
+  useTrackQuery,
+  useUserQuery,
+} from "~/graphql/gql.gen";
+import { useI18n } from "~/i18n/index";
+
+const StorySliderAction: React.FC<{ story: Story }> = ({ story }) => {
+  const { t } = useI18n();
+  const [{ data: { user } = { user: undefined } }] = useUserQuery({
+    variables: { id: story.creatorId },
+  });
+
+  return (
+    <div className="px-2 py-4 absolute z-10 bottom-0 w-full">
+      <p className="text-lg text-foreground-secondary text-center mb-1">
+        {t(story?.isLive ? "listen.promptJoin" : "listen.promptJoinNolive", {
+          username: user?.username || "",
+        })}
+      </p>
+      <Link href={`/story/${story?.id}`}>
+        <a className="btn btn-primary w-full">
+          {story?.isLive
+            ? t("listen.actionJoin")
+            : t("listen.actionJoinNoLive")}
+        </a>
+      </Link>
+    </div>
+  );
+};
 
 const StorySliderView: React.FC<{ story: Story }> = ({ story }) => {
+  const { t } = useI18n();
   const {
     state: { fetching: fetchingPlayer, crossTracks, playingStoryId },
+    skipForward,
   } = usePlayer();
 
   const [
@@ -33,13 +66,26 @@ const StorySliderView: React.FC<{ story: Story }> = ({ story }) => {
 
   return (
     <>
-      <div className="w-full mx-auto h-full box-border">
+      <div className="relative w-full h-full box-border">
         <PlayerView
           Header={<StoryNav story={story} />}
           hideControl
           track={track}
           fetching={fetching}
         />
+        {/* TODO: a11y */}
+        {/* eslint-disable-next-line jsx-a11y/click-events-have-key-events */}
+        <div
+          role="button"
+          tabIndex={0}
+          className="absolute top-24 w-full opacity-0"
+          aria-label={t("player.skipForward")}
+          onClick={skipForward}
+          style={{
+            height: "calc(100% - 14rem)",
+          }}
+        />
+        <StorySliderAction story={story} />
       </div>
     </>
   );
