+++ src/i18n/locales/en.json
@@ -179,7 +179,7 @@
     "play": "Play story",
     "feed": {
       "title": "Story feed",
-      "instruction": "Double tap to play next songs. Scroll or swipe to discover more stories."
+      "instruction": "Tap to play next songs. Scroll or swipe to discover more stories."
     }
   },
   "user": {
