+++ src/i18n/locales/vi.json
@@ -178,7 +178,7 @@
     "play": "Phát story",
     "feed": {
       "title": "Dòng story",
-      "instruction": "Chạm hai lần để phát bài tiếp theo. Cuộn chuột hoặc vuốt lên để khám phá thêm Story."
+      "instruction": "Chạm để phát bài tiếp theo. Cuộn hoặc vuốt lên để khám phá thêm Story."
     }
   },
   "user": {
